nixos/nixos-option: Summarize very large sets: Avoid printing nixpkgs
When NixOS users set the nixpkgs.pkgs option, the whole nixpkgs tree
appears in the NixOS config.  nixpkgs is very large.  Evaluating it
recursively takes more RAM than I have - at least tens of GB.

Most of the problem is the many slightly different versions of complete
copies of other parts of the tree.  For example, there might be

      14  versions of ghc
       2  Host & target
       2  Stable & unstable
* 13,877  Haskell packages
--------
 777,112  Instantiations of haskell packages

Also, most use cases of this tool probably don't want to see 56 copies
of the list of Haskell packages anyway.

After this change, nixpkgs.pkgs, if set, prints as
"«set with 10091 attributes»".