timescaledb: Fix license: asl20 -> unfree
Timescaledb can built in 'apache-licensed-code-only' mode to get
apache-licensed build artifacts.  But the nixpkgs build before this
point has been the unfree build.

Fixing this will unavoidably break builds.  We can either
 1. Default to the unfree build & break everyone not using allowUnfree, or
 2. Default to the free build & break folks using the unfree functionality

Build-unfree mode builds the "tsl" library.  This NOT NEEDED for the
postgresql extension; the timescaledb nixos test exercises the postgresql
extension and it works fine with the free apache-licensed-only build.

So, we prefer option #2 -- default to the free build -- on the basis that
 * Basic timescaledb functionality works fine in free-build-mode.
 * timescaledb should be available by default.  We don't want to push
   folks to use allowUnfree unnecessarily.

But, to help people who have (perhaps unknowingly) come to depend upon the
unfree functionality, we make this change in two steps:

Step 1. Just fix the license, marking this package unfree, but without
        changing its contents.  (This commit)

Step 2. Flip the default to apache-licensed-only.